+++ src/main/java/com/appraisaltool/model/CurrentUser.java
@@ -17,7 +17,7 @@ public class CurrentUser extends org.springframework.security.core.userdetails.U
         return user;
     }
 
-    public Long getId() {
+    public Integer getId() {
         return user.getUserId();
     }
