+++ src/main/java/com/appraisaltool/repository/AppraisalTypeRepository.java
@@ -4,7 +4,7 @@ import java.util.List;
 import org.springframework.data.jpa.repository.JpaRepository;
 import com.appraisaltool.model.AppraisalType;
 
-public interface AppraisalTypeRepository extends JpaRepository<AppraisalType, Long>{
+public interface AppraisalTypeRepository extends JpaRepository<AppraisalType, Integer>{
 
 	List<AppraisalType> findByOrderByAppraisalTypeIdAsc();
 }
