+++ src/main/java/com/appraisaltool/service/CurrentUserService.java
@@ -4,6 +4,6 @@ import com.appraisaltool.model.CurrentUser;
 
 public interface CurrentUserService {
 
-    boolean canAccessUser(CurrentUser currentUser, Long userId);
+    boolean canAccessUser(CurrentUser currentUser, Integer userId);
 
 }
